prepare for replacement
Prepared for entirely replacing the way code treat parent degrees of
freedom. Next commit should be ready to completely remove them!
